refactor: extract constraints and fields to _internal/
