+++ tests/test_integration/test_end_to_end_nested.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import Annotated
 
 from conformly import case, cases
-from conformly.constraints import GreaterOrEqual, MinLength
+from conformly._internal.constraints import GreaterOrEqual, MinLength
 
 
 @dataclass
